rename foldres and update index.ts with imports
