+++ src/BaseComponents/SortStrategyIcon/SortStrategyIcon.tsx
