+++ src/BaseComponents/SortStrategyIcon/index.ts
