+++ src/BaseComponents/styles/HeadCell.scss
