+++ src/Components/DataTableHeadCell.tsx
@@ -1,10 +1,10 @@
 import { useCallback } from "react";
-import { SORT_STRATEGY } from "../../const";
-import { HeadCell, SortStrategyIcon } from "../../Components";
-import { getColumnKey } from "../../utils";
-import { FilterContainer } from "../../FilterContainer";
-import { DataTableColumn } from "../types";
-import { HeadCellProps } from "../../Components/HeadCell/HeadCell";
+import { SORT_STRATEGY } from "../const";
+import { HeadCell, SortStrategyIcon } from "../BaseComponents";
+import { getColumnKey } from "../utils";
+import { FilterContainer } from "../FilterContainer";
+import { DataTableColumn } from "../DataTable/types";
+import { HeadCellProps } from "../BaseComponents";
 
 export type DataTableHeadCellProps = HeadCellProps & {
   column: DataTableColumn
