+++ src/Components/index.ts
@@ -1,8 +1,11 @@
-export * from "./Checkbox"
-export * from "./HeadCell"
-export * from "./SearchField"
-export * from "./TextField"
-export * from "./SortStrategyIcon"
+export * from "./SelectedCheckboxCell"
+export * from "./DataTableCell"
+export * from "./DataTableRow"
 
-export { default as Button } from "./Button"
-export type { ButtonProps } from "./Button"
+export { default as DataTableCell } from "./DataTableCell"
+export { default as DataTableHeadCell } from "./DataTableHeadCell"
+export { default as DataTableRow } from "./DataTableRow"
+
+export type { DataTableCellProps } from "./DataTableCell"
+export type { DataTableHeadCellProps } from "./DataTableHeadCell"
+export type { DataTableRowProps } from "./DataTableRow"
