--- src/DataTable/Components/index.ts
@@ -1,6 +1,0 @@
-export * from "./SelectedCheckboxCell"
-export * from "./DataTableCell"
-export * from "./DataTableRow"
-
-export { default as DataTableCell } from "./DataTableCell"
-export { default as DataTableRow } from "./DataTableRow"
