+++ src/DataTable/DataTableBody.tsx
@@ -1,8 +1,8 @@
 import { Body, Cell, CellProps, Row } from "@grossb/react-table"
 import { memo, useCallback } from "react";
 import { SELECT_ALL_STATUSES } from "../const";
-import DataTableRow from "./Components/DataTableRow";
-import useDataTableContext from "./Context/DataTableContext/useDataTableContext";
+import { DataTableRow } from "../Components";
+import { useDataTableContext } from "./Context";
 import { DataRow, DataTableColumn, RowPropsWithoutChildren } from "./types";
 
 export type DataTableBodyProps = {
