+++ src/DataTable/index.ts
@@ -1,4 +1,3 @@
-export * from "./Components"
 export * from "./Context"
 export * from "./hooks"
 export * from "./types"
